nmslib indexer moved to indexers folder
